requirements.txt
openai>=1.0.0
google-genai>=1.0.0
pyyaml>=6.0
Pillow>=10.0.0
requests>=2.31.0
